Uploaded Full Unity Project
This is my entire unity project for the game, minus a few items i'd like to use for other projects. Not sure if it works entirely but it's basically everything. Feel free to load it up and see how it works. I got stuck on optimizing the game when loading and saving chunks to file.